Originated Lecture 2
Includes several smart contracts which compile correctly.

Basic code showing 5 variable types. String, bool, int, uint, and uint8 initialized as public variables.

Second contract is an enumerator where the state, from ready, waiting, or active, can be checked. Enumerators are used for example in games to define a set without getting into variables and integers in a messy way. Eg, enum direction {North, West, East, South} and can be attached to a varaible.

Third Contract Explains "Structs" usage. Structs can be used with an array to give a variable multiple subvariables and an array can put them together. A person in this case has first and last name, but could also have soc#, age, etc. The array, Person[] has many people inside it which all have their variables first & last name. uint256 cannot go negative and is used to create a counting cache of the people so that a person who does not exist cannot be referenced.

The fourth contract explains "mapping" and is another way to use arrays for putting in data. Important note: blockchain is kind of like a big storage unit so creating data about things is an important skill.