Update with fixes and new smart contracts.
First smart contract improves on the last lecture set with a people count function. This is internal rather than public and deployment. This also adds an admin, called owner, who is the user that initializes the constructor and deploys the contract. If the person is not the owner, it will display an error and refuse to run the code, saving the gas fee.