Origination
This code is the first time experimenting with eth manipulation. The program sets a mapping for balances and an address for the payable wallet. an even for purchasing is set using the indexed buyer and amount. The constructor runs a loop for the wallet. The end results is a simple contract that can be deployed on the JS VM and buys tokens, sending eth from the buyer address to the seller address.

The second smart contract utilizes what was learned in the first contract and combines everything to create our first ERC20 Token for minting with ethereum using a smart contract.  We use  some of the code from the first smart contract and amend the buyToken bit. We create a very simple erc20 token with a mint function. You deploy the erc20 first and then the smart contract which will take in ether, mint a token, and give to the seller.