Origination - Escrow Contract - Timelock Contract
This is the origination. Adds a simple escrow account where funds can be deposited and withdrawn between two accounts.

The escrow account uses an agent address that is permitted to operate the deposit/withdraw functions. Then when it is deployed a specific address is the escrow address and the other is the payee. The funds can be transferred between.

The timelock references unix time or epoch time and compares the current block timestamp to the release time. when the condition is satisfied, the beneficiary can be paid.